+++ README.md
@@ -1,27 +1,16 @@
-# Task-2
-## Адаптировать программу, которая считает количество дней, оставшихся до нового года
-- "Форкнуть" репозиторий
-- В своем форке создать две ветки
-	- `docs` — для документации
-	- `feature` — для разработки основной функциональности
-- Перейти в ветку `feature` 
-- Выделить тесты в отдельную директорию `tests`
-- Выделить функциональную часть в директорию `src`
-- Адаптировать функцию таким образом, чтобы она считала промежуток времени между двумя указанными датами и выводила этот промежуток в человеко-читаемом виде `__ дней __ часов __ минут __ секунд __`  при этом нулевые значения с начала строки выводиться не должны
-- написать `main.js/.ts` файл, в котором вызывается написанная вами функция для вывода информации до времени оставшегося до сессии, НГ и защиты
-- адаптировать тесты
-- Не забыть адаптировать package.json: 
-run start должен запускать вашу программу run test тесты
-- слиться с веткой `feature → main`
-- перейти в ветку `docs` И написать документацию к вашему проекту, в которой следует указать
-	- Что делает ваша программа
-	- Как запускаются тесты
-	- Как запускается основная программа
-- Слить ветку `docs → main`
-- Все 3 ветки `docs` `feature` `main` залить на github
-- Запросить PR (pull request) (Только для ветки main)
+# New Year Countdown
 
-# Дополнение **Царь горы**: [rin-cast-9](//github.com/rin-cast-9)
-Убрать из соответствующего поля предыдущего участника и написать свой ник
+This program calculates the remaining time between the current date and three specific events:
+- New Year
+- End of the academic session
+- Graduation defense
 
-Пул реквесты принимаются только от **Царя горы**
+The time is displayed in a human-readable format: `__ days __ hours __ minutes __ seconds`, with zero values excluded from the output.
+
+## How to run the program
+
+To run the main program and display the time until the specified events, use the following command:
+
+```bash
+npm start
+```
