--- projects/kovalevegor/countdown.js
@@ -1,15 +1,0 @@
-function getTimeUntilNewYear(date) {
-    const nextYear = date.getFullYear() + 1;
-    const newYear = new Date(`January 1, ${nextYear} 00:00:00`);
-    const timeDiff = newYear - date;
-  
-    const days = Math.floor(timeDiff / (1000 * 60 * 60 * 24));
-    const hours = Math.floor((timeDiff % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-    const minutes = Math.floor((timeDiff % (1000 * 60 * 60)) / (1000 * 60));
-    const seconds = Math.floor((timeDiff % (1000 * 60)) / 1000);
-  
-    return { days, hours, minutes, seconds };
-  }
-  
-  module.exports = getTimeUntilNewYear;
-
